+++ changelog.md
@@ -29,7 +29,7 @@ does not on Windows.
 
 `thermos-logger` can now alternatively write the logged data into a SQLite 3
 database. The file format can be set by the new `--type` parameter. While
-`--type csv` still writes data to a CSV file, specifying '--type db` will cause
+`--type csv` still writes data to a CSV file, specifying `--type db` will cause
 the program to write the data as SQLite 3 database instead.
 
 If no type is given, then the file is assumed to be a SQLite 3 database.