+++ changelog.md
@@ -3,7 +3,8 @@
 ## Next Version (2022-05-??)
 
 `thermos-db2csv`, a command line application that converts logged data of
-`thermos-logger`from the SQLite 3 database format into the CSV format, is added.
+`thermos-logger` from the SQLite 3 database format into the CSV format, is
+added.
 
 A typo in the help text of `thermos-logger` is fixed.